* added an implementation of FFT algorithm in Memsermath * added a routine called FastLaplaceConvolution in Mesmermath for fast convolution of data, as required in MESMER ILT, and for calculation of tunneling corrections to classical k(E)s.  several tests were performed comparing the output with the standard convolution output. in the FastLaplaceConvolution output, it was found that the values of the initial elements suffer from precision issues.  For the ILT convolution, only the first few elements are unreliable, but a substantial fraction of the elements following the tunneling   convolution are unreliable.  this is due to the very small values for several elements in the tunneling probabilities vector, at energies far below the classical threshold.  thus, FastLaplaceConvolution includes a standard convolution algorithm  to fix values that are in error by more than 1%. * fixed an error in MesmerILT that convoluted the wrong data sets * increased the precision in timeEvolution routine to long double to get better species profiles at long times * updated baselines:  pentyl, HSO2, and cyclopropene are essentially unchanged, and the differences in the acetyl+O2 and acetylene + OH baselines result from the changes listed above.  with the FFT, both of these simulations are now much faster, and should be added to the standard baseline testing protocol for subsequent changes.
SVN: r164@trunk

+++ src/ReactionManager.cpp
@@ -645,7 +645,7 @@ namespace mesmer
     /* calculate the time points */
     vector<double> timePoints;
     for (int i = 0; i < maxTimeStep; ++i){
-      long double time = pow(10., static_cast<double>(i) / 10. - 11.);
+      long double time = powl(10., static_cast<double>(i) / 10. - 11.);
       if (time > maxEvoTime)
         break;
       timePoints.push_back(time);
@@ -691,7 +691,7 @@ namespace mesmer
     for (int timestep = 0; timestep < maxTimeStep; ++timestep){
       long double numColl = m_meanOmega * timePoints[timestep];
       for (int j = 0; j < smsize; ++j) {
-        work2[j] = work1[j] * exp(m_eigenvalues[j] * numColl);
+        work2[j] = work1[j] * expl(m_eigenvalues[j] * numColl); 
       } // now |wk2> = exp(Dt)*V^(T)*|init> = exp(Dt)*U^(-1)*|p(0)>
       for (int j = 0; j < smsize; ++j) {
         long double sum = 0.;
